Updates based on CR
